chore: add new landing page and move post to posts page
